Adds TLS 1.2 flag to existing value
The `SecurityProtocol` property is a flagged enum, so can have a combination of values.
https://docs.microsoft.com/en-us/dotnet/api/system.net.servicepointmanager.securityprotocol?view=netframework-4.7.2

There could be a scenario where another security protocol (e.g. SSL 3.0) may be used in as well as TLS 1.2.

This patch explicitly checks if TLS 1.2 is enabled, if not, it adds the enum flag to the exisiting value.